+++ src/main/java/com/example/demo/KeyWord/KeyWordRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.Query;
 
 import java.util.Optional;
 
-public interface KeyWordRepository extends JpaRepository<KeyWord, Long> {
+public interface KeyWordRepository extends JpaRepository<KeyWord, Integer> {
     @Query("SELECT k from KeyWord k where k.text = ?1")
     Optional<KeyWord> findKeyWordByText(String text);
 }
